Error-prone compile-type analysis (#1389)
* introduction of error-prone in gradle build; first fixes of common mistakes
* fix a few common and uncommon error-prone warnings
* move ZoneId.systemDefault() for LocalDate to a utils class

+++ core/src/main/java/org/jobrunr/utils/uuid/UUIDv7Factory.java
@@ -96,7 +96,7 @@ public final class UUIDv7Factory {
         this.incrementType = builder.getIncrementType();
         this.incrementSupplier = builder.getIncrementSupplier();
         this.random = new Random();
-        this.versionMask = (long) 7 << 12;
+        this.versionMask = 7L << 12;
 
         // initialize internal state
         this.lastUuid = make(clock.millis(), random.nextLong(), random.nextLong());
